画像抽選機能の実装
